Nickname and Role fix along with update
There was a small issue where the bot would sometimes not set a Nickname or a Role, this has been fixed by adding some time between them.

I have also added a debug toggle to help figure out where the bot might be hitting issues if you encounter any and as requested I have included the ability to add users beforehand that are to be given specific roles.